Update submodules (#34)
This PR updates the submodules to the latest commit and changes the openssl submodule to the Akamai fork (which is what is currently in PR to merge into openssl master). This openssl branch is much more up to date than the previous one and contains some fixes that were causing openssl to crash.