Fix Timer Wheel Use After Free of Connection (#39)
This PR fixes some random crashes we were seeing in the timer wheel. The problem was that a connection could sometimes get migrated to a new worker after it was uninitialized. This would cause the connection to get added to the new worker's timer wheel, and then promptly freed (returned to a lookaside list).

The fix is to make sure not to ever migrate to a new worker if the connection was already uninitialized. To ensure this, the UpdateWorker flag is cleared in the connection's uninitialize function. A few debug asserts are added to catch any possible future regressions.