Path Challenge can be limited by Path Allowance (#42)
While running the interop tool I hit a debug assert when trying to frame a path challenge. There wasn't enough room in the packet, and that was because the Path's allowance wasn't enough to fit the frame. This PR updates the code to check the limit before hand, but it is also more resilient to any other unforeseen issues as well (by cleaning up the builder).