Cross Platform PowerShell Helper Scripts (#52)
This PR adds cross platform PowerShell (6 or 7) scripts for building, testing and logging. There is still a little left to do before it can be integrated into the Azure Pipeline, but that is the ultimate goal.

One big improvement of the test.ps1 script over calling the test program directly is that it will allow for collecting logs for each individual test case separately. The down side is that it significantly increases the execution time of the tests. I'm still hoping the time can be cut down some how.