Abandon Internal Closed/Shutdown Connections (#60)
Certain test cases, when run in debug mode, where hitting an assert in the timer wheel uninitialize path because there was still a connection in the timer wheel. This ended up being because the connection was leaked. It was leaked because the connection was abandoned early on the server side, but never completely cleaned up.

This PR adds some extra logging to make it easier to diagnose these issues and fixes the problem by making sure these "internal" connections that have been closed are cleaned up immediately.