Rotate Retry Token key (#71)
This change rotates the encryption key used for protecting Stateless Retry tokens every 30 seconds.
The rotation is performed in a lazy manner, to avoid needing a timer, and only actually occurs when a Stateless Retry token is being sent from the server.
In order to prevent the encryption key from being freed while in use, all callers are required to hold a global lock while using the key. This could impact performance and can be improved by either ref-counting the keys, or by having encryption keys duplicated per CPU.
The existing tests already cover Stateless Retry tokens being sent. A test case is needed for the negative case where the client responds with the token after the key for that token has been rotated.