Connection Partitioning Refactor (#77)
This PR refactors how MsQuic does connection partitioning. It removes any logic specific to if RSS is enabled or disabled. Instead, it just has the design of keeping the connection on the primary path's receive core. So, if something changes in the system to cause the primary path's core to change, or the primary path itself changes and has a different core, then the connection will be migrated to the new core (and associated worker). To improve lookup performance, new connection IDs will be generated as well (retiring the old ones) and sent to the peer to use.