Allow API Calls on unstarted Stream (#145)
This PR refactors the stream logic to allow send and shutdown to be called before start, so that these operations can be batched up and then triggered all at once when finally started. It also has the effect of allowing StreamShutdown to be called on a stream that isn't yet started or failed to be started (flow control blocked).

Generally, this relaxes the logic required for the app to synchronize things when calling the stream.