Add support for building in UWP mode
Flags needed to be added to the build. Right now we have to force override a compiler defined macro in the platform header file. Looking to see if there is a way to do this only through the command line.

Running tests is possible, but requires installing the msix file, so not extremely feasable. I'll look into this more as well. I did add a build however so we can be sure this at leasts continues to build