Implementing more functionality in the macOS datapath
There's a lot of sockopts that .. failed to work in macOS. Also a core
issue due to how QUIC_ADDR is setup. In Linux, sockaddr_in's first
member is the si_family; on macOS, that's not the case, so the union of
sockaddr, sockaddr6, and si_family is harmful here.

Also there's some noisey printfs across the code to help me track down
why connections get closed before they even get opened. They'll be
removed before the final PR.