Wait for all connections to be freed during registration close (#474)
* Wait for all connections to be freed during registration close

If the connection received data after the registration worker pools were free, we hit a crash. This is caused by a race between cleaning up a connection and the binding receiving data. If we wait for the connections to be fully clean before cleaning up the worker threads, the race is avoided.

Should fix #472