Move from si_family in the union to sockaddr;
This way we don't rely on the internal sockaddr structure ordering. On
darwin for example, si_family is not the first member. There's no
guarantee this structure will stay the same, and the proper way to do
this is to just use a sockaddr as the first member, which is guaranteed
to be the first member of both sockaddr_in and sockaddr_in6.