Properly pend packets that couldn't be send due to lack of buffer space
I think the empty-pending-queue function is still a bit funky, and might
cause the CPU to spin excessively if the kernel can't get packets out
fast enough. Also, once in a while I get ECONNREFUSED, but and I'm not
sure whether or not to try to send the packet again, or to go to the
Unreach handler. It happens nondeterministically enough that I think
it's just that one of the machines in the network is bloated, but it's
actually possible the machine just isn't available.