Build in an msquic directory
SourceLink is a relative path, relative to the source directory. If 2 completely unrelated projects build on AZP, they'll likely get the same base source path, causing them to have the same relative path root. Its not really specified what happens if 2 pdbs with source link with the same roots are debugged in a process. I'm asking, but might as well play it safe and build in an msquic pathed source directory