Add RPS Latency metrics to performance testing (#931)
RPS latency numbers are recorded for every RPS run, as there is no noticeable impact on performance for doing so. 

These latency numbers are currently not uploaded to the Database, they will be added at a later date. They are printed though.

Histograms of latency percentiles are also generated. Currently only the numbers from the last run of a set are uploaded to artifacts. 

A 60 connection test was added to reduce utilization to a constant level.