A few more WPA code quality cleanups
It's much better to suppress the compiler styles we don't want in the csproj rather then the editorconfig, as that way the project stays independent of the repo. Just a bit more future proof.

Also fixes a few missing readonly variables, and 1 missing public Debug.Assert.

Also converts some things public things that could be IReadOnlyLists to IReadOnlyLists.