macOS Support (#1226)
Adds macos support to msquic.

Currently, it builds, and core tests will pass, but many datapath tests will not. This is because the datapath depends on ipv4 and ipv6 dual mode sockets, which have issues on macos. However, we want this merged so it doesn't get 4 months stale again.

The datapath is basically a clone of the epoll datapath, except using kqueue.

Co-authored-by: Max <mxms@me.com>
Co-authored-by: Nick Banks <nibanks@microsoft.com>
Co-authored-by: Lars Eggert <200328+larseggert@users.noreply.github.com>