Build requires cmake 3.19
On Windows (although the issue is not platform specific), when the `openssl` Tls option is chosen. CMakeLists takes the path that one of the `PROPERTY FOLDER` lines are touched. CMAKE versions under 3.19 doesn't allow FOLDER property to be used.

Repro Steps : Try `scripts/build.ps1 -Tls openssl` with cmake 3.18.xx