Add script to merge macos binaries (#1298)
Apple recommends shipping universal binaries on macOS, and actually require it if shipping to store. The easiest way to do this is post build. This adds a script that will merge all artifact binaries that can then be used and shipped.

Not currently hooked into azure, as this requires we build both platforms with the same tls provider, and arm64 openssl builds have issues there. I did test this locally, and got universal binaries that worked.