Inline allocate event objects on posix (#1299)
* Inline allocate event objects on posix

Previously, each event required an allocation. This was because the API for events required an opaque object. Use a macro to work around this limitation, and inline allocate the events

* Actually properly use variable

* Fix event addressof

* Inline function calls as well

* Fix inline declarations

* 1 more fix

* Revert changes to make all event function calls inline

For some reason, it seems to TANK perf.

* Force a cxplat_event to align to 16 byte boundaries

* Switch things back to inline

* Move all events to the top of their structs.

This will make sure they are both aligned correctly, and don't cause extra padding