Properly detect arm current platform for macos pwsh builds (#1388)
Theres no direct flag in powershell to detect which architecture is being used. Use uname to get the processor, and if it shows up as x86_64 check to see if we're running in translated mode.

Works on an M1 mac. CI as that is running native on Intel, that will test the native intel case. Native arm powershell does not exist yet, but when it does that case is handled as well