Fix generator changes not working for clean openssl builds (#1419)
In CI, we use build caching to speed up the build. This made the OpenSSL pass in CI, however in a clean build the wrong version of OpenSSL would get built.

Fix up the build so this continues to work. Also fixes a missed PGO variable. Doesn't affect CI builds, but would affect local builds.