Switch from _SC_NPROCESSORS_ONLN to _SC_NPROCESSORS_CONF on posix platforms.
On posix systems, CONF is the number of CPUs available in the system, whereas ONLN is the number of CPUs available to the scheduler. However, it looks like its trivally possible to change the 2nd number with a few config knobs, and could even happen behind the scenes. We should be using the max possible to ensure that we never have less partitions then are actually available.