Always treat macOS as a single core platform (#1427)
* Always treat macOS as a single core platform

On M1 macs, there is no way to get the current CPU. So our only option is to treat the system as a single core.
On Intel macs, with certain configurations CPUID can return processor numbers that are greater then the number the syscall tells us. This results in assertions, or out of range calls in release mode. We don't currently have a workaround, so treat Intel macOS as a single core.

Since both platforms have current core issues, TL;DR just always treat macOS as a single core platform.