Always send data through the datapath thread (#1448)
By doing this, the worker thread will get more time to handle encryption and decryption, as the datapath thread will handle the send. This looks to be about 8-10% throughput improvement with no regressions elsewhere.

Only works on Windows User mode currently. Epoll in theory should get similar improvements, but there are some other issues there. kqueue and windows kernel mode this change is not applicable to.