Add stats logging option to perf, log debug output to file
Logging connection stats can be an easy way to debug certain issues with the performance tool. Add an option to make this work.

By default, the pwsh script only prints logs in debug mode. Instead, write all output to a file as well, so this can be later reviewed as well. Otherwise all data from step 1 would just be lost.