Add stats logging option to perf, log debug output to file (#1471)
* Add stats logging option to perf, log debug output to file

Logging connection stats can be an easy way to debug certain issues with the performance tool. Add an option to make this work.

By default, the pwsh script only prints logs in debug mode. Instead, write all output to a file as well, so this can be later reviewed as well. Otherwise all data from step 1 would just be lost.

* Fix builds

* Spacing nit

* Removing extra comma

* Fix incorrect log file

* Extra backtick

* 1 more parameter check

* Actually log local results

* Add build version to machine name

* Fix log name

* Hopefully 1 last fix

* Don't overwrite release logs with debug logs

* Toggle kernel as well

Co-authored-by: Nick Banks <nibanks@microsoft.com>