Reduce number of loops in RandomParitionId tests.
We run a LOT of test scenarios, and on unix platforms this test takes multiple minutes. On windows it takes > 10 seconds, which is still a bit crazy. This reduces the number of iterations. With how many times we run this test, the amount of randomness we are testing is still high.