If queued sends are enabled, ALWAYS send on datapath
In a high throughput scenario, if the send is alternating between the datapath and quic worker thread, the amount of packet reording can be massive. It can even be high enough to cause the connection to be killed. To make this work correctly, we need to have a flag for if any sends on the socket are queued, we need to queue until empty. However, this likely will be a larger change, and increase contention, so revert for now to remove the functional breakage.