Use the endpoints of the current key phase to decide old or new encryption keys
A bug was found while working on a separate PR that would cause the packets to be decoded incorrect upon key update. The root cause ended up being our method for choosing old or new keys when the opposite key phase was detected was incorrect. This PR updates the logic for selecting the keys to match 6.5 of the QUIC TLS spec.

The correct method uses the endpoints of the packets received on the current key phase to decide which keys to use.