Properly update the start point of the current key phase to decide old or new encryption keys (#1548)
* Use the endpoints of the current key phase to decide old or new encryption keys

A bug was found while working on a separate PR that would cause the packets to be decoded incorrect upon key update. The root cause ended up being our method for choosing old or new keys when the opposite key phase was detected was incorrect. This PR updates the logic for selecting the keys to match 6.5 of the QUIC TLS spec.

The correct method uses the endpoints of the packets received on the current key phase to decide which keys to use.

* A few failure fixes, along with a spelling comment

* Fix key update

* Add telemetry asserts

* Remove unable to decrypt check

* Drop packets that would use old keys if we have no old keys

* Remove telemetry assert

* Don't do anything on the middle type packet

* Revert "Drop packets that would use old keys if we have no old keys"

This reverts commit 76cb64079d07997f3511c9d84c2a6432c8786d2d.

* Only log if lower range is triggered

* Only use the start of the range, also update test

* Fix some comments

* Add comments for test update

* Add random loss key update tests

* Missing break

* Formatting

* Fix double new line