Fix compilation on newer windows SDKs
Newer SDKs define some of the APIs we use to set thread information that we used to define locally. However, these APIs are not gated behind any version macros in the SDK headers, so we can't just conditionally check for the existance of them in the preprocessor. The solution is to rename the local versions of the struct and enum definition to not colide. Additionally, NtSetInformationThread needed its signature slightly changed to be compatible with the SDK