Fix queuing 0-RTT data during handshake
Previously, if 0-RTT data was queued during the handshake, it has the potential to not queue the flush send. This makes it so if we have 0-RTT flags and sends are queued, we'll allow them to be flushed.

Also, there are cases where we could end up with stream data but no flags, which if this occured during the handshake would hit an assert. This assert is only there to tell if we're going to do no work, but this would only be a slight perf penalty when it does happen, so its not a valid assert