Fix wraparound bug in CxPlatSocketPoolGetNextLocalPort (#2105)
Reset NextLocalPort back to the (currently hardcoded) start port when it wraps.
Otherwise we'll use port 0, which is invalid, or a low reserved port number.
CxPlatSocketPoolGetNextLocalPort is also renamed to CxPlatSockPoolGetNextLocalPort
for consistency with sibling functions.

Synchronization is changed from interlocked to use of the RW lock. On this note,
probably we should make a future change to select the port in CxPlatTryAddSocket
while we already hold the lock; that function is called right after this one currently.