Add support for building arm64ec binaries (#2146)
* Add support for building arm64ec binaries

Arm64ec is an emulation compatible arm64 calling convention. It allows x64 and arm64 code to mix in the same exectable. This is great for a program that uses plugins, as it can use exisiting x64 plugins but run most of its code as arm code.

Only supports schannel, as openssl would require changes to their assembly files, which they do not have any interest in doing.

* Formatting nit

* Use windows 2022