Avoid fetching OpenSSL digests and ciphers
The old cipher returning calls like EVP_aes_128_gcm() perform late binding
which means they fetch on init.  Fetch in OpenSSL 3.0 is a relatively
expensive operation.  Instead of fetching every time a cipher is required,
it is faster to pre-fetch and reuse the same EVP_CIPHER object.

Likewise, HMAC is better prefetched but it has the additional complexity
of fetching a digest internally.  Instead of just prefetching the
EVP_MAC object, it is better to create an EVP_MAC_CTX object and call
EVP_MAC_CTX_dup() as required.

According to the profiling I've done, this represents a circa 4% boost
in HPS.