Enable many more tests for XDP (#2572)
Now that the unit tests are enabled on the XDP platform, this PR enables most of the main
tests. There are still six sets of tests that have failures and will need to be fixed and enabled,
but to split up the work these tests are disabled in this PR by adding them to the filter defined
in the XDP test job in azure-pipelines.ci.yml.