Resolve route in QuicSendFlush and enable LoLa testcases that are fixed by this change (#2618)
There are cases where QuicSendFlush is called without queueing the flush with QuicSendQueueFlush, and cases where the path is switched between the route resolution in QuicSendQueueFlush and the actual flush happening. In both cases the result can be that we try to send a packet on an unresolved route. This is fixed by moving route resolution to QuicSendFlush from QuicSendQueueFlush.

QuicConnProcessRouteCompletion has also been changed to call QuicSendFlush instead of QuicSendQueueFlush. Since the main point of QuicConnProcessRouteCompletion is to finally do a send after the route resolution is finished, there isn't any point in going back to the end of the queue again.