Fix QuicTestCibirExtension for LoLa and enable *CibirExtension* testcases (#2622)
When the server uses CIBIR but the client doesn't, the connection should be rejected.
In regular sockets, the client's invalid handshake makes it up to QUIC and a transport
error is sent. But for LoLa, there is a filter based on the CIBIR ID for packet delivery to
the server, so the client's connect attempt times out instead. We already have a check
at the end for the connection not being successfully set up, so this can be addressed
by simply not asserting that we don't time out waiting for the handshake to complete.