Spinquic job for LoLa (#2623)
Spinquic revealed a synchronization bug: during cleanup, the datapath may be cleaned up before a half-opened connection, which results in a crash when trying to free pending RX packets in the connection's ReceiveQueue. This is fiexed by adding a CXPLAT_DATAPATH::SocketsRundown similar to the one in winuser for the raw datapath.