Remove GetIfEntry2 call in route resolution (#2768)
Calling GetIfEntry2 to get mac address for the selected interface in route resolution is not necessary. Modifying MAC address isn't even supported on all NICs and it requires a NIC restart on the ones that support it AFAIK. We can just query MAC address once during interface initialization (this has been already implemented but the PhysicalAddress field isn't used) and use it in route resolution.

It did make xperf traces look much prettier as GetIfEntry2 is more expensive (sending an OID through every layer in the NDIS stack) than the sum of all other iphlpapi calls but I don't see a noticeable difference in HPS.

The next bottleneck in HPS test is schannel RPCs from the xperf traces I collected.