Optimize QuicStreamSetUpdateMaxStreams (#4852)
Store new streams that are blocked by Stream Id flow control in a sorted list (instead of the hash table containing all streams) to improve the performances of QuicStreamSetUpdateMaxStreams.

It will now be enough to iterate over a subset of this list instead of iterating over the entirety of opened streams when receiving a "MAX_STREAMS" frame increasing the limit.