Support use of old differently named artifacts
Since the test-down-level test uses artifacts constructed with the old
openssl/quictls naming convention, we need to inform the test of what
those old names are.  Introduce the imgtls element to the matrix vectors
to store those names.

Note, to avoid having to do lots of conditional checking, just create
imgtls names for each vector, and make it a 1:1 mapping for schannel.