Prevent duplicate definitions of CXPLAT_DBG_ASSERT (#5016)
CXPLAT_DBG_ASSERT is failing compilation in the fuzzer here:
https://github.com/google/oss-fuzz/actions/runs/14454075706/job/40663846040?pr=13222

Because both msquic.hpp and quic_platform.h define the macro, and only
the former has ifndef guards around the definition.  This is not
problematic in other inclusion sites because quic_platform.h is always
included first.  This is not true in the case of fuzz.cc, which reverses
the inclusion order.

Correct fuzz.cc to include files in the same order as other inclusion
sites to avoid the duplicate definition issues

Co-authored-by: QUIC Dev[bot] <quicdev@microsoft.com>