Back out openssl support for linux package building
The current package building scheme on linux configures for the use of
the system openssl libcrypto.

However, using a newer libssl statically with the system libcrypto
(which may be from an older version), while it might work, isn't
actually supported by openssl, as the newer libssl might reference
functions that don't exist in libcrypto.

It works the other way around (i.e. an older libssl will work with a
newer libcrypto) from the same major stream, as all the functions
refrenced by the older libssl are guaranteed to be present in the newer
libcrypto (which is why quictls works with this setup), but it will
never be guaranteed to work when the locally built static libssl is a
newer version than the systems libcrypto.so.

So for now don't build openssl backing with this feature enabled.

It should probably be considered to remove this feature entirely, as its
almost guaranteed that, if you keep the local openssl submodule up to
date, it will always be newer than what distributions are shipping.