Remove incorrect assertion on determining Key for Control frames
We're seeing periodic failures here:
https://github.com/microsoft/msquic/actions/runs/14776565585/job/41488589785?pr=4959

Due to an attempt to ACK packets for which we have a read key, but not a
write key.

However, based on discussion here:
https://github.com/microsoft/msquic/pull/4959#issuecomment-2845296376

I believe its a possibly expected scenario (based on the TLS backend)
for this to be ok.  Given that, and the fact that callers of
QuicPacketBuilderGetPacketTypeAndKeyForControlFrames are written to
expect a failure case gracefully, that its ok to remove this assertion