Add 1RTT secrets into TlsSecrets for openssl (#5222)
* Add 1RTT secrets into TlsSecrets for openssl

When doing the initial authoring, I neglected to add the 1RTT secrets
into the TlsSecrets struct, meaning the SSLKEYLOG file wasn't populated
with those values.  Fix that up

Fixes #5218

* fixup! Add 1RTT secrets into TlsSecrets for openssl

* fixup! Add 1RTT secrets into TlsSecrets for openssl

* Update src/platform/tls_openssl.c

Co-authored-by: Guillaume Hetier <hetier.guillaume@gmail.com>

* Make sure UniquePtrs get destructed after MsQuicRegistration

We were failing one of the DataTests due to the Client/Server secrets
getting destroyed before the underlying TLS structure was done with
them.

This was occuring because in QuicTestConnectAndPing, the test function
returns before the openssl TLS layer had completing registering the
server side application read key, which isn't needed for the reporting
of an established connection.  The function returned, cleaned the
secrets buffer, after which the TLS layer tried to access them, and
boom.

Thanks to @guhetier, who noted that the destruction of the
MsQuicRegistration forces shutdown of the connection, and blocks on its
completion, so if we move the declaration of the UniquePtr buffers to
before the allocation of said registration, the fact that C++ calls
destructors in reverse order ensures that the Secrets memory remains
valid until the connection is shut down.

---------

Co-authored-by: Guillaume Hetier <hetier.guillaume@gmail.com>