[CP] Fix double deref in connection pool error path (#5597) (#5601)
## Description

When creating a connection pool, if a `QuicConnStart` fails:
- the connection was marked as `ExternalOwner` to prevent it from
sending notification to the app
- but this also mean that the closing logic will take care of releasing
the owner refcount, since the application is not the owner yet
- the connection was closed using `MsQuicConnectionClose`, which release
the refcount of the application

This caused a double release, triggering an assertion.

We should not call APIs from internal call (it makes logging confusing
and breaks some assumptions), so queue the connection close manually
instead.

Fixes #5550.

## Testing

C/I.
Need to consider if there is a simple way to deterministically test the
connection pool failure paths.