Build for Android Arm/Arm64 in CI (#5620)
Arm and Arm64 android builds feel much more applicable than the x86_64
android builds that CI already produces. Since it was a simple change to
make to the CI workflow, I figured I'd send the proposal to add them to
the CI flow along with the PR to actually add it. It's up to you whether
the extra CI work makes sense, though.