Fix validateGlobalParam test on ADO pipelines (#5600)
## Description

The `validateGlobalParam` test has been failing on the ADO pipeline.
This seems to be due to the `QuicTestValidateRegistration` test now
creating registrations for real, which triggers the lazy initialization
of the Quic library and creates a datapath.
These API tests assume nothing has been initialized and test for failure
path.

Instead, move the test to the "basic test" file.

As a follow up, we should consider making test truly independent by not
creating the MsQuic library globally.