Fix the Key update test (#5604)
## Description

The "KeyUpdate" test is sometime failing spuriously.

This was due to a sequence of frames where, after a client initiated key
update, it would not need to send a frame in response to the server
acknowledging the key update. Then, the server would still be waiting
for the key update confirmation when the test tries to initiate a key
update from the server side.
This behavior follows the RFC, there is no need to acknowledge a key
change a packet needs to be sent for another reason.

The test was relying on the assumption that no frames would be sent
without a test action, which is not true in general (only most of the
time). It was also mixing two largely independent test scenarios in one
test, making it hard to understand.

This PR:
- split the test in two based on the two scenarios validated (force
updates / trigger updates by sending packets)
- refactor the test, adding helpers, removing non-needed scopes
- fix the issue by ensuring both the client and the server send a frame
to acknowledge a key update.

Fixes #3463.