Fail test passes when a test succeeds only after a retry (#5636)
## Description

Retries have been added automatically to every test on failure, which
might be hiding infrequent issues.
This change makes the test pass fail if it takes a retry for the CI to
succeed.

This will let us know how frequently the retry was needed to avoid
spurious issues and how often it was hiding real infrequent issues. If
the failure rate does not dramatically increase, the entire retry logic
should be removed.