Fix some flakyness in QuicTestOperationPriority (#5640)
## Description

`QuicTestOperationPriority` relied on the number of stream bytes sent
from the statistic data to determine if a send operation completed or
not.
This is incorrect, as the number of bytes is updated in the stats only
after a `FLUSH_SEND` operation. It left room for some situations where
the `GET_PARAM` operation could run between the `STREAM_SEND` and the
`FLUSH_SEND`, causing a spurious test failure.

The test now relies on the `STREAM_SEND_COMPLETE` indication.

## Testing

CI. Ran the test 100 times locally.

## Documentation

N/A. Comment added to the test code.