Add logging for spinquic tests. (#5643)
## Description

The Spin powershell script had support for gathering MsQuic logs but the
CI is never using it.

Let's add an option to configure logging in spin tests. For workflow
dispatch, the user can control the log profile. For pull request and
merge events, Basic.Light is used.

## Testing

CI

## Documentation

N/A