Fix invalid handle access for RegistrationCloseCleanup Events (#5673)
## Description

When MsQuic is first initialized, there may not be enough memory to
allocate the datapath layer.

This happens so rarely that there is little to no coverage for error
handling when this happens.

But when it happens, you skip directly to the exit condition, and that
tries free-ing memory that was never initialized.

So add a flag to cleanup only when the events are properly initialized. 

## Testing

Stress tests.

## Documentation

No